Map a memory object when calling data() if it isn't already mapped.
This fixes a segault when try to map output tensors/images which were
not given initial data.

Add tests for copying to an uninitialised tensor and image to the
OpTensor/ImageCopy test suites.

Move mapRawData and unmapRawData functions into the Memory class
instead of duplicating them in Tensor and Image.

Move mPhysicalDevice, mDevice, mPrimaryMemory and mStaging memory into
the Memory class instead of duplicating them in Tensor and Image.