Map memory if it is not currently mapped before returning it.
Add tests for reading output Tensors/Images after running an Algorithm.
The data()/vector() functions for these were not mapping the memory
before returning the pointers.

Also add a warning if eStorage Tensors or Images are initialsed with
data. This is no longer required now that output Tensors/Images are
supported.