Create the staging image as linear tiled.
This fixes a problem with mapping the staging memory to the host when it
has been tiled in some (potentially unknown) way. It needs to be mapped
linearly so that the host program can update it.

Let the primary memory still be optimally tiled.

Add a 2D image copy test which shows this issue (at least on PowerVR).