Add a new MemoryType eDeviceAndHost
This memory is device memory which is also host accessible and host
coherent. This memory type is supported for both images and tensors.

For images, add support for selecting the tiling mode. It can be optimal
or linear. The default is optimal tiled.

Linear is required for any host accessible memory (becuase
mapping a tiled image doesn't make sense when you don't know how it's
been tiled). Staging memory for eDevice memory is always linearly tiled
for this reason.