Fix incorrect test for invalid tiling.
Linear tiling can be used on images of any memory type.
Optimal tiling can only be used on images of type eDevice or eStorage.

Add overloads to Manager::image and Manager::imageT to make specifying a
tiling format optional. If no tiling format is specified then it will be
inferred from the memory type. Optimal tiling will be used for images of
type eDevice and eStore. Linear tiling will be used for images of type
eHost and eDeviceAndHost.

Add an init function for Image which does the work of the constructor.
All the constructors call this. This is required to allow the Image
constructor which doesn't have a tiling format to infer the tiling
format before calling the init function (which used the be the Image
constructor). You can't call a constructor from another constructor
after running code, so using a function for common code gets the
behaviour we want.